Predictor/requirements.txt
 -f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.0+cpu
torchvision==0.9.0+cpu
Flask==1.0.3
requests==2.25.1
Pillow==8.2.0
gunicorn